+++ src/Sample.cpp
@@ -180,7 +180,7 @@ SampleCache* Sample::getOrCreateCache(SampleHolder* sampleHolder, int32_t phaseI
 	}
 
 
-	uint32_t keyWords[4];
+	uintptr_t keyWords[4];
 	keyWords[0] = phaseIncrement;
 	keyWords[1] = timeStretchRatio;
 	keyWords[2] = skipSamplesAtStart;
